+++ custom_components/ovum_mira/ovum_mira_modbus/__init__.py
@@ -1,8 +1,13 @@
 from .config import InstallationOptions
 from .device import OvumMiraSystem
 from .enums import (
+    BufferLoadingStatus,
     BufferSystemType,
+    CoolingBufferLoadingStatus,
+    DhwRequestStatus,
+    FreshWaterDrawStatus,
     HeatingCircuitMode,
+    HeatingCircuitTargetMode,
     HeatingCircuitType,
     PvStatus,
     SwitchState,
@@ -10,8 +15,13 @@ from .enums import (
 )
 
 __all__ = [
+    "BufferLoadingStatus",
     "BufferSystemType",
+    "CoolingBufferLoadingStatus",
+    "DhwRequestStatus",
+    "FreshWaterDrawStatus",
     "HeatingCircuitMode",
+    "HeatingCircuitTargetMode",
     "HeatingCircuitType",
     "InstallationOptions",
     "OvumMiraSystem",